1.8.8 - Product Delete Check
Added a function comment and check to make sure a product exists before deleting it. This isn't important now due to the context of where it is called, however it would be important in the future if the application was scaled and this function where to be called out of context.